+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Pretty print JSON error response
   * perf: front-load HTML template and stylesheet at middleware construction
+  * perf: only load template and stylesheet once
   * perf: resolve file paths at start up
 
 1.4.3 / 2016-01-17